fix(rebase): correct errors from rebase
