+++ docs/components/menu/index.html
@@ -34,37 +34,7 @@ assets:
     </hx-menu>
   {% endcode %}
 </section>
-
-<section>
-<h2 class="hxSectionTitle" id="split-menu">Split Menu</h2>
-  <div class="demo">
-    <div class="hxBtnGroup">
-      <button class="hxBtn hxBtn--primary">Split Menu</button>
-      <hx-disclosure class="hxBtn hxBtn--primary" aria-controls="splitmenuId">
-        <hx-icon class="menu-icon" type="angle-down"></hx-icon>
-      </hx-disclosure>
-    </div>
-    <hx-menu class="hxBox-sm" id="splitmenuId" >
-      <hx-menuitem class="hxBtn hxBtn--link">Action 1</hx-menuitem>
-      <hx-menuitem class="hxBtn hxBtn--link">Action 2</hx-menuitem>
-      <hx-menuitem class="hxBtn hxBtn--link">Action 3</hx-menuitem>
-    </hx-menu>
-  </div>
-  {% code 'html' %}
-    <div class="hxBtnGroup">
-      <button class="hxBtn hxBtn--primary">Split Menu</button>
-      <hx-disclosure class="hxBtn hxBtn--primary" aria-controls="splitmenuId">
-        <hx-icon class="menu-icon" type="angle-down"></hx-icon>
-      </hx-disclosure>
-    </div>
-    <hx-menu class="hxBox-sm" id="splitmenuId" >
-      <hx-menuitem class="hxBtn hxBtn--link">Action 1</hx-menuitem>
-      <hx-menuitem class="hxBtn hxBtn--link">Action 2</hx-menuitem>
-      <hx-menuitem class="hxBtn hxBtn--link">Action 3</hx-menuitem>
-    </hx-menu>
-  {% endcode %}
-</section>
-
+<!--
 <section>
 <h2 class="hxSectionTitle" id="cog-menu">Cog Menu</h2>
   <div class="demo">
@@ -88,6 +58,7 @@ assets:
     </hx-menu>
   {% endcode %}
 </section>
+-->
 
 <section>
   <h2 class="hxSectionTitle" id="advanced-menu">Advanced Menu</h2>
@@ -131,6 +102,36 @@ assets:
         <hx-menuitem class="hxBtn hxBtn--link">Action 2</hx-menuitem>
         <hx-menuitem class="hxBtn hxBtn--link">Action 3</hx-menuitem>
       </hx-menuitem-group>
+    </hx-menu>
+  {% endcode %}
+</section>
+
+<section>
+<h2 class="hxSectionTitle" id="split-menu">Split Menu</h2>
+  <div class="demo">
+    <div class="hxBtnGroup" id="split-group">
+      <button class="hxBtn hxBtn--primary">Split Menu</button>
+      <hx-disclosure class="hxBtn hxBtn--primary" aria-controls="splitmenuId">
+        <hx-icon class="menu-icon" type="angle-down"></hx-icon>
+      </hx-disclosure>
+    </div>
+    <hx-menu class="hxBox-sm" id="splitmenuId" relative-to="split-group">
+      <hx-menuitem class="hxBtn hxBtn--link">Action 1</hx-menuitem>
+      <hx-menuitem class="hxBtn hxBtn--link">Action 2</hx-menuitem>
+      <hx-menuitem class="hxBtn hxBtn--link">Action 3</hx-menuitem>
+    </hx-menu>
+  </div>
+  {% code 'html' %}
+    <div class="hxBtnGroup">
+      <button class="hxBtn hxBtn--primary">Split Menu</button>
+      <hx-disclosure class="hxBtn hxBtn--primary" aria-controls="splitmenuId">
+        <hx-icon class="menu-icon" type="angle-down"></hx-icon>
+      </hx-disclosure>
+    </div>
+    <hx-menu class="hxBox-sm" id="splitmenuId" >
+      <hx-menuitem class="hxBtn hxBtn--link">Action 1</hx-menuitem>
+      <hx-menuitem class="hxBtn hxBtn--link">Action 2</hx-menuitem>
+      <hx-menuitem class="hxBtn hxBtn--link">Action 3</hx-menuitem>
     </hx-menu>
   {% endcode %}
 </section>
