+++ src/helix-ui.less
@@ -153,4 +153,5 @@ fieldset label {
 @import 'alert'; // needs turned into core element + theme
 @import 'tabset'; // needs rewritten as several custom elements
 @import 'navigation'; // component and theme need separated
+@import 'menu'; // likely needs refactored
 /* ====== end:OVERRIDES ========== */
