+++ src/helix-ui/elements.js
@@ -1,5 +1,6 @@
 export { HXCheckboxElement } from './elements/HXCheckboxElement.js';
 export { HXDisclosureElement } from './elements/HXDisclosureElement.js';
+export { HXMenuElement } from './elements/HXMenuElement.js';
 export { HXIconElement } from './elements/HXIconElement.js';
 export { HXRevealElement } from './elements/HXRevealElement.js'; //
 export { HXTabpanelElement } from './elements/HXTabpanelElement.js'; //
