+++ src/helix-ui/styles/menu.less
@@ -8,10 +8,12 @@ hx-disclosure {
 
 hx-menu {
   background-color: @gray-0;
+  border: 1px solid @gray-300;
   box-shadow: 0px 3px 3px 0 @shadow-color;
   display: none;
+  max-width: 20rem;
+  min-width: 10rem;
   position: absolute;
-  width: 15rem;
   z-index: @menu-z-index;
 
   &[open] {
